3-langchain/ask-my-invoice/requirements.txt
langchain
gradio
python-dotenv
langchain-google-genai
pypdf
chromadb
langchain-community
langchain-chroma
numpy